rename AppConfigDriver to ConfigurationDriver, allow customizing nuget global folder
