+++ Reqnroll.TestProjectGenerator/Reqnroll.TestProjectGenerator/ConfigurationDriver.cs
@@ -1,0 +1,29 @@
+using System;
+using System.Configuration;
+
+namespace Reqnroll.TestProjectGenerator;
+
+public class ConfigurationDriver
+{
+    public string TestProjectFolderName => GetConfigSetting("testProjectFolder", "RR");
+    public string VSTestPath => GetConfigSetting("vstestPath", "Common7\\IDE\\CommonExtensions\\Microsoft\\TestWindow");
+    public string MsBuildPath => GetConfigSetting(nameof(MsBuildPath));
+    public bool PipelineMode => GetConfigSwitch(nameof(PipelineMode));
+    public string GlobalNuGetPackages => GetConfigSetting(nameof(GlobalNuGetPackages));
+
+    public bool GetConfigSwitch(string key, bool defaultValue = false)
+    {
+        return GetConfigSetting(key, defaultValue.ToString()).Equals("true", StringComparison.InvariantCultureIgnoreCase);
+    }
+
+    public string GetConfigSetting(string key, string defaultValue = null)
+    {
+        var envSetting = Environment.GetEnvironmentVariable($"REQNROLL_TEST_{key}".ToUpperInvariant());
+        if (!string.IsNullOrEmpty(envSetting)) return envSetting;
+
+        var configSetting = ConfigurationManager.AppSettings[key];
+        if (!string.IsNullOrEmpty(configSetting)) return configSetting;
+
+        return defaultValue;
+    }
+}
