+++ Reqnroll.TestProjectGenerator/Reqnroll.TestProjectGenerator/ProjectBuilder.cs
@@ -214,7 +214,7 @@ namespace Reqnroll.TestProjectGenerator
 
             _project = new Project(ProjectName, ProjectGuid, Language, TargetFramework, Format, ProjectType);
 
-            _testProjectFolders.PathToNuGetPackages = _project.ProjectFormat == ProjectFormat.Old ? Path.Combine(_testProjectFolders.PathToSolutionDirectory, "packages") : _folders.RunUniqueGlobalPackages;
+            _testProjectFolders.PathToNuGetPackages = _project.ProjectFormat == ProjectFormat.Old ? Path.Combine(_testProjectFolders.PathToSolutionDirectory, "packages") : _folders.GlobalNuGetPackages;
             if (!Directory.Exists(_testProjectFolders.PathToNuGetPackages))
             {
                 Directory.CreateDirectory(_testProjectFolders.PathToNuGetPackages);
